fixed deletion bug that resulted in deleting complete pipeline steps even if "skip already processed" was enabled

+++ src/components/Documents/ConceptClusterDialog.vue
@@ -276,7 +276,7 @@ function statusToString(status?: PipelineResponseStatus) {
 }
 
 function statusDetailsToString(steps?: ConceptGraphPipelineStatus[]) {
-  let returnString = 'Unfinished steps: '
+  let returnString = t('conceptCluster.unfinishedSteps') + ': '
   steps
     ?.filter((step: ConceptGraphPipelineStatus) => {
       return step.status != ConceptGraphPipelineStatusEnum.Finished
@@ -355,9 +355,13 @@ function confirmStartPipeline() {
       message: t('conceptCluster.confirmStart')
     }
   }).onOk(() => {
-    deletePipeline()
-      .then(startPipeline)
-      .catch((e: Error) => renderError(e))
+    if (graphPipeline.value?.steps?.length != undefined
+      && graphPipeline.value?.steps.length < 4
+      && !skipPresent.value) {
+      deletePipeline()
+        .then(startPipeline)
+        .catch((e: Error) => renderError(e))
+    } else { startPipeline().catch((e: Error) => renderError(e)) }
   })
 }
 
@@ -415,7 +419,7 @@ async function startPipeline() {
       .then(loadGraphPipeline)
   } else {
     return conceptPipelineApi
-      ?.startConceptGraphPipeline(props.dataSource.id, props.dataSource.id, undefined, undefined, language.value)
+      ?.startConceptGraphPipeline(props.dataSource.id, props.dataSource.id, skipPresent.value, false, language.value)
       .then(pipelineResponseCheck)
       .then(loadGraphPipeline)
   }
